Escapadebars
Escapade has been the premium events bar operator in the North West and further afield for over a decade. We create unforgettable memories for guests through our drinks and service providing all the necessary elements required to curate a seamless event experience, including a wide selection of cocktails, refreshments, staffing, and end-to-end solutions.